+++ libraries/session/src/main/java/androidx/media3/session/MediaUtils.java
@@ -244,7 +244,8 @@ import java.util.List;
       long lastSetPlayWhenReadyCalledTimeMs,
       long timeDiffMs) {
     boolean receivedUpdatedPositionInfo =
-        lastSetPlayWhenReadyCalledTimeMs < playerInfo.sessionPositionInfo.eventTimeMs;
+        playerInfo.sessionPositionInfo.equals(SessionPositionInfo.DEFAULT)
+            || (lastSetPlayWhenReadyCalledTimeMs < playerInfo.sessionPositionInfo.eventTimeMs);
     if (!playerInfo.isPlaying) {
       if (receivedUpdatedPositionInfo || currentPositionMs == C.TIME_UNSET) {
         return playerInfo.sessionPositionInfo.positionInfo.positionMs;