+++ RELEASENOTES.md
@@ -49,6 +49,9 @@
 *   Muxers:
 *   IMA extension:
 *   Session:
+    *   Fix issue where the current position jumps back when the controller
+        replaces the current item
+        ([#951](https://github.com/androidx/media/issues/951)).
 *   UI:
     *   Fallback to include audio track language name if `Locale` cannot
         identify a display name